backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend service"
authors = ["joshhn <duy.huynhngoc.05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
bs4 = "^0.0.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pylint = "^3.3.1"
pre-commit = "^3.8.0"
pyright = "^1.1.382.post0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint]
extension-pkg-whitelist="pydantic" # Pylint cannot lint C extension modules https://stackoverflow.com/a/67560637

[tool.pyright]
exclude = [
    "**/__pycache__",
]
reportOptionalMemberAccess = false
reportGeneralTypeIssues = false

[tool.pydocstyle]
ignore = [
    "D104", # Docstring in public package.
    "D107", # Doctstring in __init__.
    "D203", # One blank line required before class docstring.
    "D212", # Multi-line docstring summary should start at the first line.
]
